Added Expense List
Trying to separate concerns with an Operations/Storage/View layer

+++ app/src/main/java/com/vijaysharma/expenses/database/models/Expense.java
@@ -6,7 +6,7 @@ public class Expense {
     public Long _id;
     public String description;
     public String comment;
-    public float amount;
+    public double amount;
     public Date date;
     public String serverId;
 }
